+++ test/cli/array.js
@@ -2,7 +2,7 @@ const { chaca, schemas } = require("../../lib/cjs/index.js");
 
 module.exports = chaca.schema({
   id: chaca.key(chaca.sequence()),
-  username: schemas.internet.userName(),
+  username: schemas.internet.username(),
   image: { type: schemas.image.people(), possibleNull: 50 },
   email: schemas.internet.email(),
   posts: { type: schemas.id.uuid(), isArray: { min: 1, max: 3 } },