+++ test/schemas/internet/username.test.ts
@@ -1,28 +1,28 @@
 import { schemas } from "../../../src";
 
-describe("# Internet userName test", () => {
+describe("# Internet username test", () => {
   it("Without arguments", () => {
-    const username = schemas.internet.userName().getValue();
+    const username = schemas.internet.username().getValue();
     expect(typeof username).toBe("string");
   });
 
   it("Pass firstName argument", () => {
     const username = schemas.internet
-      .userName()
+      .username()
       .getValue({ firstName: "hector" });
     expect(username.includes("hector")).toBe(true);
   });
 
   it("Pass lastName argument", () => {
     const username = schemas.internet
-      .userName()
+      .username()
       .getValue({ lastName: "hector" });
     expect(username.includes("hector")).toBe(true);
   });
 
   it("Pass firstName and lastName argument", () => {
     const username = schemas.internet
-      .userName()
+      .username()
       .getValue({ lastName: "hector", firstName: "john" });
     expect(username.includes("hector") && username.includes("john")).toBe(true);
   });